refactor: split id_tracker_base.rs into a module
Break the 501-line file into a directory module with logically separated
components: trait_def.rs (IdTracker trait + constants), point_mappings_ref.rs
(PointMappingsRefEnum + PointMappingsGuard), tracker_enum.rs (IdTrackerEnum +
its IdTracker impl), and mod.rs (re-exports). No logic changes.
