+++ lib/segment/src/id_tracker/id_tracker_base/mod.rs
@@ -1,0 +1,7 @@
+mod point_mappings_ref;
+mod tracker_enum;
+mod trait_def;
+
+pub use point_mappings_ref::{PointMappingsGuard, PointMappingsRefEnum};
+pub use tracker_enum::IdTrackerEnum;
+pub use trait_def::{DELETED_POINT_VERSION, IdTracker};
